# producer-service/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Kafka Producer Demo",
    "description": "Kafka producer API to injest test order messages.",
    "version": "0.0.1-SNAPSHOT"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/v1/orders": {
      "post": {
        "tags": [
          "order-controller"
        ],
        "summary": "Post multiple orders to Kafka for downstream processing",
        "operationId": "publishRecords",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CanonicalOrder"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultsResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CanonicalOrder": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "minLength": 1
          },
          "customerId": {
            "type": "string",
            "minLength": 1
          },
          "orderTimestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "minLength": 1,
            "pattern": "PENDING|PROCESSING|SHIPPED|CANCELLED"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            },
            "minItems": 1
          },
          "shippingAddress": {
            "$ref": "#/components/schemas/ShippingAddress"
          },
          "version": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          }
        },
        "required": [
          "orderTimestamp",
          "shippingAddress"
        ]
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "sku": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1
          },
          "quantity": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "unitPrice": {
            "type": "number",
            "format": "double",
            "minimum": 0.01
          }
        }
      },
      "ShippingAddress": {
        "type": "object",
        "properties": {
          "street": {
            "type": "string",
            "minLength": 1
          },
          "city": {
            "type": "string",
            "minLength": 1
          },
          "postalCode": {
            "type": "string",
            "minLength": 1
          },
          "country": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "ResultsResponse": {
        "type": "object",
        "properties": {
          "processed": {
            "type": "integer",
            "format": "int32"
          },
          "successful": {
            "type": "integer",
            "format": "int32"
          },
          "failed": {
            "type": "integer",
            "format": "int32"
          },
          "failedOrderIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}